// server/src/movies/dto/create-movie.dto.ts
export class CreateMovieDto {
    _id: string;
    title: string;
    overview: string;
    poster_path: string;
    backdrop_path: string;
    release_date: string;
    original_language: string;
    tagline: string;
    genres: string;
    casts: string;
    vote_average: string;
    runtime: string;
}